Aparentemente los unsigned long son solo de 4 bytes en windows mientras que en linux son de 8. Hay que cambiar el tipo de variable por algo más estable

+++ include/chip8.h
@@ -42,7 +42,7 @@ class Chip8{
             0xF0, 0x80, 0xF0, 0x80, 0xF0, //E
             0xF0, 0x80, 0xF0, 0x80, 0x80  //F
         };
-        unsigned long screen[32]; // 8 bytes
+        unsigned long long screen[32]; // 8 bytes
        
         
         void (Chip8::*Chip8Table[16]) () = {
